Update README.md
This repository contains 25 website projects that I created. Each project includes a README file that describes the project and provides instructions for setting up and running the project. The projects are organized into categories, and each category has a README file that describes the category. Each project also includes a link to the GitHub repository where the project is hosted.